int64: AllSimdHelper totalSize and loop counters to long
- AllSimdHelper<T> parameter: int totalSize → long totalSize
- Loop counters and vectorEnd: int → long
- Part of int64 indexing migration

+++ src/NumSharp.Core/Backends/Kernels/ILKernelGenerator.Reduction.Boolean.cs
@@ -54,7 +54,7 @@ namespace NumSharp.Backends.Kernels
         /// SIMD helper for All reduction with early-exit.
         /// Returns true if ALL elements are non-zero.
         /// </summary>
-        internal static unsafe bool AllSimdHelper<T>(void* input, int totalSize) where T : unmanaged
+        internal static unsafe bool AllSimdHelper<T>(void* input, long totalSize) where T : unmanaged
         {
             if (totalSize == 0)
                 return true; // NumPy: all([]) == True (vacuous truth)
@@ -64,9 +64,9 @@ namespace NumSharp.Backends.Kernels
             if (Vector256.IsHardwareAccelerated && Vector256<T>.IsSupported && totalSize >= Vector256<T>.Count)
             {
                 int vectorCount = Vector256<T>.Count;
-                int vectorEnd = totalSize - vectorCount;
+                long vectorEnd = totalSize - vectorCount;
                 var zero = Vector256<T>.Zero;
-                int i = 0;
+                long i = 0;
 
                 // SIMD loop with early exit
                 for (; i <= vectorEnd; i += vectorCount)
@@ -91,9 +91,9 @@ namespace NumSharp.Backends.Kernels
             else if (Vector128.IsHardwareAccelerated && Vector128<T>.IsSupported && totalSize >= Vector128<T>.Count)
             {
                 int vectorCount = Vector128<T>.Count;
-                int vectorEnd = totalSize - vectorCount;
+                long vectorEnd = totalSize - vectorCount;
                 var zero = Vector128<T>.Zero;
-                int i = 0;
+                long i = 0;
 
                 for (; i <= vectorEnd; i += vectorCount)
                 {
@@ -115,7 +115,7 @@ namespace NumSharp.Backends.Kernels
             else
             {
                 // Scalar fallback
-                for (int i = 0; i < totalSize; i++)
+                for (long i = 0; i < totalSize; i++)
                 {
                     if (System.Collections.Generic.EqualityComparer<T>.Default.Equals(src[i], default))
                         return false;
